requirements.txt
Django>=3.2.2,<3.3.0
djangorestframework>=3.12.4,<3.13.0
drf-access-policy>=0.9.0,<0.9.1
psycopg2>=2.8.6,<2.9
